requirements.txt
langchain
streamlit
pypdf
faiss-cpu
langchain-community
unstructured
markdown
sentence-transformers
langchain-huggingface
python-dotenv
langchain-google-genai
